Refactor Vulkan rendering components
- Removed the RenderLoop class and integrated its functionality into VulkanRenderLoop.
- Cleaned up unused imports and optimized code structure for clarity.
